Re-organize component namespace
