+++ resources/js/components/icons/StarIcon.vue
